+++ apps/cowswap-frontend/src/common/pure/Modal/index.tsx
@@ -5,6 +5,7 @@ import { isMobile } from '@cowprotocol/common-utils'
 import { Command } from '@cowprotocol/types'
 import { Media, UI } from '@cowprotocol/ui'
 
+import { t } from '@lingui/core/macro'
 import { useSpringValue, useTransition } from '@react-spring/web'
 import { useGesture } from '@use-gesture/react'
 import styled from 'styled-components/macro'
@@ -31,7 +32,7 @@ interface ModalProps {
  */
 // TODO: Break down this large function into smaller functions
 // TODO: Add proper return type annotation
-// eslint-disable-next-line max-lines-per-function, @typescript-eslint/explicit-function-return-type
+// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function Modal({
   isOpen,
   onDismiss,
@@ -87,7 +88,7 @@ export function Modal({
                       style: { transform: y.interpolate((y) => `translateY(${(y as number) > 0 ? y : 0}px)`) },
                     }
                   : {})}
-                aria-label="dialog content"
+                aria-label={t`dialog content`}
                 $minHeight={minHeight}
                 $maxHeight={maxHeight}
                 $mobile={isMobile}