+++ apps/cowswap-frontend/src/modules/erc20Approve/containers/TradeApproveModal/useTradeApproveCallback.ts
@@ -15,50 +15,12 @@ import { useHandleApprovalError } from './useHandleApprovalError'
 import { useApproveCallback } from '../../hooks'
 import { useResetApproveProgressModalState, useUpdateApproveProgressModalState } from '../../state'
 
-interface ProcessTransactionConfirmationParams {
-  response: TransactionResponse
-  currency: Currency | undefined
-  account: string | undefined
-  spender: string | undefined
-  chainId: number | undefined
-  setOptimisticAllowance: (data: { tokenAddress: string; owner: string; spender: string; amount: bigint; blockNumber: number; chainId: number }) => void
-}
-
-async function processTransactionConfirmation({
-  response,
-  currency,
-  account,
-  spender,
-  chainId,
-  setOptimisticAllowance,
-}: ProcessTransactionConfirmationParams): Promise<TradeApproveResult<TransactionReceipt>> {
-  const txResponse = await response.wait()
-
-  if (!chainId) {
-    return { txResponse, approvedAmount: undefined }
-  }
-
-  const approvedAmount = processApprovalTransaction(
-    {
-      currency,
-      account,
-      spender,
-      chainId,
-    },
-    txResponse,
-  )
-
-  if (approvedAmount) {
-    setOptimisticAllowance(approvedAmount)
-  }
-
-  return { txResponse, approvedAmount: approvedAmount?.amount }
-}
-
 interface TradeApproveCallbackParams {
   useModals: boolean
   waitForTxConfirmation?: boolean
 }
+
+const EVM_TX_HASH_LENGTH = 64 + 2
 
 const DEFAULT_APPROVE_PARAMS: TradeApproveCallbackParams = {
   useModals: true,
@@ -121,6 +83,12 @@ export function useTradeApproveCallback(currency: Currency | undefined): TradeAp
 
         approvalAnalytics('Sign', symbol)
 
+        // Check the length to skip waiting for Safe tx
+        // We have to return undefined in order to avoid jumping into confirm screen after approval tx sending
+        if (response.hash.length !== EVM_TX_HASH_LENGTH) {
+          return undefined
+        }
+
         if (waitForTxConfirmation) {
           return await processTransactionConfirmation({
             response,
@@ -160,4 +128,51 @@ export function useTradeApproveCallback(currency: Currency | undefined): TradeAp
       handleApprovalError,
     ],
   ) as TradeApproveCallback
+}
+
+interface ProcessTransactionConfirmationParams {
+  response: TransactionResponse
+  currency: Currency | undefined
+  account: string | undefined
+  spender: string | undefined
+  chainId: number | undefined
+  setOptimisticAllowance: (data: {
+    tokenAddress: string
+    owner: string
+    spender: string
+    amount: bigint
+    blockNumber: number
+    chainId: number
+  }) => void
+}
+
+async function processTransactionConfirmation({
+  response,
+  currency,
+  account,
+  spender,
+  chainId,
+  setOptimisticAllowance,
+}: ProcessTransactionConfirmationParams): Promise<TradeApproveResult<TransactionReceipt>> {
+  const txResponse = await response.wait()
+
+  if (!chainId) {
+    return { txResponse, approvedAmount: undefined }
+  }
+
+  const approvedAmount = processApprovalTransaction(
+    {
+      currency,
+      account,
+      spender,
+      chainId,
+    },
+    txResponse,
+  )
+
+  if (approvedAmount) {
+    setOptimisticAllowance(approvedAmount)
+  }
+
+  return { txResponse, approvedAmount: approvedAmount?.amount }
 }