+++ apps/cowswap-frontend/src/modules/ordersTable/pure/OrderStatusBox/getOrderStatusTitleAndColor.ts
@@ -1,19 +1,22 @@
 import { UI } from '@cowprotocol/ui'
 
+import { i18n, MessageDescriptor } from '@lingui/core'
+import { msg, t } from '@lingui/core/macro'
+
 import { OrderStatus } from 'legacy/state/orders/actions'
 
 import { getIsFinalizedOrder } from 'utils/orderUtils/getIsFinalizedOrder'
 import { ParsedOrder } from 'utils/orderUtils/parseOrder'
 
-const orderStatusTitleMap: { [key in OrderStatus]: string } = {
-  [OrderStatus.PENDING]: 'Open',
-  [OrderStatus.PRESIGNATURE_PENDING]: 'Signing',
-  [OrderStatus.FULFILLED]: 'Filled',
-  [OrderStatus.EXPIRED]: 'Expired',
-  [OrderStatus.CANCELLED]: 'Cancelled',
-  [OrderStatus.CREATING]: 'Creating',
-  [OrderStatus.FAILED]: 'Failed',
-  [OrderStatus.SCHEDULED]: 'Scheduled',
+const titleMap: { [key in OrderStatus]: MessageDescriptor } = {
+  [OrderStatus.PENDING]: msg`Open`,
+  [OrderStatus.PRESIGNATURE_PENDING]: msg`Signing`,
+  [OrderStatus.FULFILLED]: msg`Filled`,
+  [OrderStatus.EXPIRED]: msg`Expired`,
+  [OrderStatus.CANCELLED]: msg`Cancelled`,
+  [OrderStatus.CREATING]: msg`Creating`,
+  [OrderStatus.FAILED]: msg`Failed`,
+  [OrderStatus.SCHEDULED]: msg`Scheduled`,
 }
 
 // TODO: Reduce function complexity by extracting logic
@@ -23,7 +26,7 @@ export function getOrderStatusTitleAndColor(order: ParsedOrder): { title: string
   // For this reason we use the flag to override the order status
   if (order.executionData.fullyFilled || order.status === OrderStatus.FULFILLED) {
     return {
-      title: orderStatusTitleMap[OrderStatus.FULFILLED],
+      title: i18n._(titleMap[OrderStatus.FULFILLED]),
       color: `var(${UI.COLOR_SUCCESS_TEXT})`,
       background: `var(${UI.COLOR_SUCCESS_BG})`,
     }
@@ -33,14 +36,14 @@ export function getOrderStatusTitleAndColor(order: ParsedOrder): { title: string
     // Partially filled is also not a real status
     if (order.executionData.partiallyFilled) {
       return {
-        title: 'Partially Filled',
+        title: t`Partially Filled`,
         color: `var(${UI.COLOR_SUCCESS_TEXT})`,
         background: `var(${UI.COLOR_SUCCESS_BG})`,
       }
     }
 
     return {
-      title: orderStatusTitleMap[order.status],
+      title: i18n._(titleMap[order.status]),
       color: order.status === OrderStatus.EXPIRED ? `var(${UI.COLOR_ALERT_TEXT})` : `var(${UI.COLOR_DANGER_TEXT})`,
       background: order.status === OrderStatus.EXPIRED ? `var(${UI.COLOR_ALERT_BG})` : `var(${UI.COLOR_DANGER_BG})`,
     }
@@ -49,7 +52,7 @@ export function getOrderStatusTitleAndColor(order: ParsedOrder): { title: string
   // Cancelling is not a real order status
   if (order.isCancelling) {
     return {
-      title: 'Cancelling...',
+      title: t`Cancelling...`,
       color: `var(${UI.COLOR_DANGER_TEXT})`,
       background: `var(${UI.COLOR_DANGER_BG})`,
     }
@@ -58,7 +61,7 @@ export function getOrderStatusTitleAndColor(order: ParsedOrder): { title: string
   // Handle signing state
   if (order.status === OrderStatus.PRESIGNATURE_PENDING) {
     return {
-      title: orderStatusTitleMap[OrderStatus.PRESIGNATURE_PENDING],
+      title: i18n._(titleMap[OrderStatus.PRESIGNATURE_PENDING]),
       color: `var(${UI.COLOR_ALERT_TEXT})`,
       background: `var(${UI.COLOR_ALERT_BG})`,
     }
@@ -67,7 +70,7 @@ export function getOrderStatusTitleAndColor(order: ParsedOrder): { title: string
   // Handle unfillable orders
   if (order.isUnfillable) {
     return {
-      title: 'Unfillable',
+      title: t`Unfillable`,
       color: `var(${UI.COLOR_DANGER_TEXT})`,
       background: `var(${UI.COLOR_DANGER_BG})`,
     }
@@ -75,7 +78,7 @@ export function getOrderStatusTitleAndColor(order: ParsedOrder): { title: string
 
   // Finally, map order status to their display version
   return {
-    title: orderStatusTitleMap[order.status],
+    title: i18n._(titleMap[order.status]),
     color: order.status === OrderStatus.PENDING ? `var(${UI.COLOR_TEXT})` : `var(${UI.COLOR_TEXT})`,
     background:
       order.status === OrderStatus.PENDING ? `var(${UI.COLOR_TEXT_OPACITY_10})` : `var(${UI.COLOR_TEXT_OPACITY_10})`,