+++ apps/cowswap-frontend/src/modules/trade/containers/QuotePolingProgress/index.tsx
@@ -3,6 +3,8 @@ import { ReactNode } from 'react'
 import { displayTime } from '@cowprotocol/common-utils'
 import { CircleProgress, HoverTooltip, UI } from '@cowprotocol/ui'
 
+import { t } from '@lingui/core/macro'
+
 import { useTradeQuoteCounter, QUOTE_POLLING_INTERVAL, useTradeQuote } from 'modules/tradeQuote'
 
 const size = 18
@@ -11,10 +13,9 @@ export function QuotePolingProgress(): ReactNode {
   const { isLoading } = useTradeQuote()
   const counter = useTradeQuoteCounter()
   const percent = Math.ceil((counter * 100) / QUOTE_POLLING_INTERVAL)
+  const time = displayTime(counter, true)
 
-  const content = (
-    <span>{counter === 0 ? 'Quote is updating...' : <>Quote will be updated in {displayTime(counter, true)}</>}</span>
-  )
+  const content = <span>{counter === 0 ? t`Quote is updating...` : t`Quote will be updated in ${time}`}</span>
 
   return (
     <HoverTooltip wrapInContainer placement="top" content={content}>