+++ apps/cowswap-frontend/src/modules/tradeQuote/types.ts
@@ -11,3 +11,8 @@ export interface TradeQuotePollingParameters {
   isQuoteUpdatePossible: boolean
   useSuggestedSlippageApi: boolean
 }
+
+export interface QuotePollingUpdateTimings {
+  now: number
+  ref: { current: number | null }
+}