+++ apps/cowswap-frontend/src/modules/wallet/containers/WalletModal/index.tsx
@@ -18,7 +18,7 @@ import { toggleAccountSelectorModalAtom } from '../AccountSelectorModal/state'
 
 // TODO: Break down this large function into smaller functions
 // TODO: Add proper return type annotation
-// eslint-disable-next-line max-lines-per-function, @typescript-eslint/explicit-function-return-type
+// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export function WalletModal() {
   const dispatch = useAppDispatch()
   const { account } = useWalletInfo()